Add Android.bp for bpftool
Build bpftool for host so we can use it to generate BPF skeleton
headers when building libbpf-tools.

For now, set -DBPFTOOL_WITHOUT_SKELETONS to avoid a multi-step build
process involving an intermediate "bootstrap" bpftool used to embed
BPF skeletons in the bpftool binary itself to support newer features
not needed for this use case.

Test: m bpftool
Bug: 203823368
Signed-off-by: Connor O'Brien <connoro@google.com>
Change-Id: I8b4e2056dfb18761ac775819c30f9bd5f199db04